+++ Source/Core/Visualization/Renderer/ParallelCoordinatesRenderer.cpp
@@ -16,24 +16,12 @@
 namespace kvs
 {
 
-/*===========================================================================*/
-/**
- *  @brief  Constructs a new ParallelCoordinatesRenderer class.
- */
-/*===========================================================================*/
-ParallelCoordinatesRenderer::ParallelCoordinatesRenderer():
-//    m_top_margin( 20 ),
-//    m_bottom_margin( 20 ),
-//    m_left_margin( 30 ),
-//    m_right_margin( 30 ),
-    m_enable_anti_aliasing( false ),
-    m_enable_multisample_anti_aliasing( false ),
-    m_active_axis( 0 ),
-    m_line_opacity( 255 ),
-    m_line_width( 1.0f ),
-    m_color_map( 256 )
+using ThisClass = ParallelCoordinatesRenderer;
+
+void ThisClass::setAntiAliasingEnabled( const bool aa, const bool msaa ) const
 {
-    m_color_map.create();
+    m_enable_anti_aliasing = aa;
+    m_enable_multisample_anti_aliasing = msaa;
 }
 
 /*===========================================================================*/
@@ -42,7 +30,7 @@ ParallelCoordinatesRenderer::ParallelCoordinatesRenderer():
  *  @param  multisample [in] if true, multisampling is available
  */
 /*===========================================================================*/
-void ParallelCoordinatesRenderer::enableAntiAliasing( const bool multisample ) const
+void ThisClass::enableAntiAliasing( const bool multisample ) const
 {
     m_enable_anti_aliasing = true;
     m_enable_multisample_anti_aliasing = multisample;
@@ -53,7 +41,7 @@ void ParallelCoordinatesRenderer::enableAntiAliasing( const bool multisample ) c
  *  @brief  Disables anti-aliasing mode.
  */
 /*===========================================================================*/
-void ParallelCoordinatesRenderer::disableAntiAliasing() const
+void ThisClass::disableAntiAliasing() const
 {
     m_enable_anti_aliasing = false;
     m_enable_multisample_anti_aliasing = false;
@@ -61,23 +49,24 @@ void ParallelCoordinatesRenderer::disableAntiAliasing() const
 
 /*===========================================================================*/
 /**
- *  @brief  Render parallel coordinates.
- *  @param  object [in] pointer to object
- *  @param  camera [in] pointer to camera
- *  @param  light [in] pointer to light
+ *  @brief  Updates color map range.
+ *  @param  table [in] pointer to the table data
  */
 /*===========================================================================*/
-void ParallelCoordinatesRenderer::exec( kvs::ObjectBase* object, kvs::Camera* camera, kvs::Light* light )
+void ThisClass::updateColorMapRange( const kvs::TableObject* table )
 {
-    kvs::IgnoreUnusedVariable( light );
+    const auto color_axis_min_value = static_cast<float>( table->minValue( m_active_axis ) );
+    const auto color_axis_max_value = static_cast<float>( table->maxValue( m_active_axis ) );
+    m_color_map.setRange( color_axis_min_value, color_axis_max_value );
+}
 
-    kvs::TableObject* table = kvs::TableObject::DownCast( object );
-
-    BaseClass::startTimer();
-
-    kvs::OpenGL::WithPushedAttrib attrib( GL_ALL_ATTRIB_BITS );
-
-    // Anti-aliasing.
+/*===========================================================================*/
+/**
+ *  @brief  Updates anti-aliasing.
+ */
+/*===========================================================================*/
+void ThisClass::updateAntiAliasing()
+{
     if ( m_enable_anti_aliasing )
     {
 #if defined ( GL_MULTISAMPLE )
@@ -96,51 +85,74 @@ void ParallelCoordinatesRenderer::exec( kvs::ObjectBase* object, kvs::Camera* ca
             kvs::OpenGL::Hint( GL_LINE_SMOOTH_HINT, GL_NICEST );
         }
     }
+}
 
+/*===========================================================================*/
+/**
+ *  @brief  Render parallel coordinates.
+ *  @param  object [in] pointer to object
+ *  @param  camera [in] pointer to camera
+ *  @param  light [in] pointer to light
+ */
+/*===========================================================================*/
+void ThisClass::exec( kvs::ObjectBase* object, kvs::Camera* camera, kvs::Light* light )
+{
+    kvs::IgnoreUnusedVariable( light );
+
+    const auto* table = kvs::TableObject::DownCast( object );
+
+    BaseClass::startTimer();
+
+    const float dpr = camera->devicePixelRatio();
+    const int x0 = m_margins.left();
+    const int x1 = camera->windowWidth() - m_margins.right();
+    const int y0 = m_margins.top();
+    const int y1 = camera->windowHeight() - m_margins.bottom();
+
+    kvs::OpenGL::WithPushedAttrib attrib( GL_ALL_ATTRIB_BITS );
     kvs::OpenGL::Enable( GL_BLEND );
     kvs::OpenGL::SetBlendFunc( GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA );
+
+    this->updateColorMapRange( table );
+    this->updateAntiAliasing();
+
+    const size_t naxes = table->numberOfColumns();
+    const float dx = float( x1 - x0 ) / ( naxes - 1 );
+
+    auto y_normalize = [&] ( double v, double v_min, double v_max )
+    {
+        return y1 - ( y1 - y0 ) * ( v - v_min ) / ( v_max - v_min );
+    };
+
+    auto y_value = [&] ( const size_t axis, const size_t row )
+    {
+        const kvs::Real64 min_value = table->minValue( axis );
+        const kvs::Real64 max_value = table->maxValue( axis );
+        const kvs::Real64 value = table->column( axis )[row].to<kvs::Real64>();
+        return y_normalize( value, min_value, max_value );
+    };
 
     kvs::OpenGL::Render2D render( kvs::OpenGL::Viewport() );
     render.begin();
     {
-        const float color_axis_min_value = static_cast<float>( table->minValue( m_active_axis ) );
-        const float color_axis_max_value = static_cast<float>( table->maxValue( m_active_axis ) );
-        const kvs::AnyValueArray& color_axis_values = table->column( m_active_axis );
-        m_color_map.setRange( color_axis_min_value, color_axis_max_value );
-
-        const float dpr = camera->devicePixelRatio();
-//        const int x0 = m_left_margin;
-//        const int x1 = camera->windowWidth() - m_right_margin;
-//        const int y0 = m_top_margin;
-//        const int y1 = camera->windowHeight() - m_bottom_margin;
-        const int x0 = m_margins.left();
-        const int x1 = camera->windowWidth() - m_margins.right();
-        const int y0 = m_margins.top();
-        const int y1 = camera->windowHeight() - m_margins.bottom();
-
+        const auto& color_axis_values = table->column( m_active_axis );
         const size_t nrows = table->column(0).size();
-        const size_t naxes = table->numberOfColumns();
-        const float stride = float( x1 - x0 ) / ( naxes - 1 );
-        for ( size_t i = 0; i < nrows; i++ )
+        for ( size_t i = 0; i < nrows; ++i )
         {
             if ( !table->insideRange( i ) ) continue;
 
             kvs::OpenGL::SetLineWidth( m_line_width * dpr );
             kvs::OpenGL::Begin( GL_LINE_STRIP );
             {
-                const kvs::Real64 color_value = color_axis_values[i].to<kvs::Real64>();
-                const kvs::RGBColor color = m_color_map.at( static_cast<float>( color_value ) );
-                kvs::OpenGL::Color( color.r(), color.g(), color.b(), m_line_opacity );
-                for ( size_t j = 0; j < naxes; j++ )
-                {
-                    const kvs::Real64 min_value = table->minValue(j);
-                    const kvs::Real64 max_value = table->maxValue(j);
-                    const kvs::Real64 value = table->column(j)[i].to<kvs::Real64>();
+                const auto color_value = color_axis_values[i].to<kvs::Real64>();
+                const auto color = m_color_map.at( static_cast<float>( color_value ) );
+                kvs::OpenGL::Color( kvs::RGBAColor( color, m_line_opacity ) );
 
-//                    const kvs::Real64 x = m_left_margin + stride * j;
-                    const kvs::Real64 x = m_margins.left() + stride * j;
-                    const kvs::Real64 y = y1 - ( y1 - y0 ) * ( value - min_value ) / ( max_value - min_value );
-                    kvs::OpenGL::Vertex( kvs::Vec2( x, y ) * dpr );
+                auto x = static_cast<kvs::Real64>( m_margins.left() );
+                for ( size_t j = 0; j < naxes; ++j, x += dx )
+                {
+                    const auto p = kvs::Vec2( x, y_value( j, i ) );
+                    kvs::OpenGL::Vertex( p * dpr );
                 }
             }
             kvs::OpenGL::End();